feat: update mantine modals for metric query
